gamesave: perform needed seek for some D2X-XL levels
The D2 level format has a header, which contains offsets for each data
array (e.g. segments, walls, lights, etc.). At some point in its
history, the d2x project removed code that seeks to these offsets, and
just assumed these arrays are located sequentially in the file. There
are assert()s that check that the offsets are really correct. They don't
trigger on any official levels are D2 extension levels. I don't know why
they did this absurd bullshit, but it's possible there's an obscure
level somewhere that was corrupted, and for which the modifications
worked. Still:

          WHAT THE FUCK

Anyway. Some D2X-XL levels break this assumption - which was completely
fine with the original game code, but breaks with code based on later
d2x, such as the code in this repository. One of them is level 3 from
the mission "The Sphere". Add the minimum of code that loads it
correctly (well, at least it looks like it might be fine?). The Assert()
is still run for non-d2x-xl levels, just so I'll be aware if a D2 level
has the same issue.