Cleanup matcen initialization
The initialization of matcens and fuelcens was unnecessarily convoluted.
For example, RobotCenters is initialized twice - once with dummy data,
and then overwritten with data from the level file. There was some
unneeded code duplication.

Remove the duplication and unnecessary code. Rename segment.value to
fuelcen_num - in theory, this field could be reused for other segment
types, but it isn't and won't, and "value" sounds far too generic.

The organization of the data is unfortunate. The segment.matcen_num
could be discarded if FuelCenter had a matcen_num field, likewise
segment.fuelcen_num is probably not really needed (strangely).
FuelCenter is dumped into savegame, so it can't be easily changed, and
also a lot of code references matcen_num. It all sucks.

matcen_info.segnum is not used and thus renamed to dummy_segnum. Not
removing it since matcen_info is also directly read/written from
savegames and level data files.

Add a larger comment about the utterly confusing way how mtcens and
segments are associated. The D2X-XL source code calls Parallax "dumb
amateurs" in this context. It also complains about inconsistencies
in official level data. I'm not sure what this refers to. The bogus
fields were already overwritten/recomputed/unused in the original D2
source code dump, so it's clear that the values of the fields as found
in level data is probably garbage. What the code produces on loading
seems to be fine. (For example, it would be a bad breakage if a level
had multiple segments that used the same matcen.)