cntrlcen: do not kill controlcenter if boss bot present in d2x-xl mode
No idea what this code is for. The game doesn't blow up just because
both bosses and controlcenters are present, and there's much much worse
that could exist in level data to break the game. So why this special
case? Nobody knows.

Level 2 in "The Sphere" has multiple bosses and a controlcenter. Do not
let it replace the latter with OBJ_GHOST. It seems D2X-XL removed this
code completely, but it does a lot of other weird stuff, and I'm not
sure if it could cause regressions with any existing levels, so do it
only in d2x-xl mode.