cntrlcen, multi: remove gun pos precomputation
The code assumes that there is a most 1 reactor. It precomputed gun
points/directions based on the absolute position and orientation of the
reactor. Obviously this broke if there were multiple reactors ("The
Sphere", secret level, reachable through level 3).

Get rid of the precomputed gun positions, and recompute them on need.
This is only when it decides to fire, which is every other second. (Or,
possibly, every frame when the player is too far away. But this could be
optimized if it really mattered.)

With this change applied, different reactors fire properly at you.

Multiplayer untested.

Full support for multiple reactors still requires per-reactor runtime
state. This would be messy (even if you managed to stuff it into struct
object, you'd need to deal with old savegames), so this commit doesn't
deal with this yet.