cntrlcen: render fireballs on all destroyed reactors
Since classic D2 has only at most 1 reactor per level, they made it a
global variable for whether to render small explosions ("fireballs") on
a destroyed reactor. Obviously this doesn't work if there are several
reactors, and it'll render them only on the reactor that was destroyed
last.

Move the fireball creation to the per-frame handler, and hack it back to
calling it for destroyed reactors (explode_object() unsets control_type,
so we just restore it).