Remove Dead_controlcen_object_num variable
With the previous commit, this variable is unused. Remove it, keep
savegames compatible by writing 0 and discarding the read value.

(Since the last commit, loading old savegames where the reactor was
destroyed won't show the "fireballs" correctly, but that's a minor
visual regression in a very obscure case that I consider acceptable.)