switch: debounce TF_PERMANENT
D2X-XL has TF_PERMANENT switches, which don't get destroyed when they're
hit. Instead they're retriggering the action. If you hit the switch with
a quad laser, the switch might be toggled up to 4 times at once. Often
this action is toggling something, so if the number of hits is even, it
will appear as if nothing is happening.

Fix this by debouncing the hit by introducing a small timeout until the
switch can be hit again. I thought about using the last_operated field,
but it's actually set only once (and unset only in obscure situations),
so just add another field in a dumb way. I don't want to touch the
idiotic D2X-XL state machine stuff anyway. There's no need to save this
new field in savegames, at worst it might glitch a little if you save in
exactly the same moment when lasers are hitting a switch.