switch: primitive D2X-XL speedboost support
This is another Descent 3 thing the D2X-XL author tried to reimplement
in D2. These are areas that accelerate the player into a certain
direction.

Unfortunately, the D2X-XL design/implementation is messy. Speedboost is
activated by a fly-through trigger, and ends when you hit another
trigger that disables it. This requires tracking for each accelerated
object if and how it is accelerated. (A simpler implementation would be
to define acceleration magnitude/direction per segment, so that all
objects within a segment are accelerated.) Since in the current state of
the code you can't just add new object fields, this gets annoying.

Another problem is that there is actually no guarantee the speedboost is
ever stopped correctly. A level can just not define a stop trigger, or
it could "miss" some possible paths in the level geometry. In fact, "The
Sphere" level 4 has a "tunnel" with gaps in the sides, so it might be
possible in theory to escape.

If the object gets moved away by other means, such as respawning or
teleportation, explicitly resetting the speedboost is required.

Yet another tricky issue is that the speedboost area isn't necessarily
entered in direction of the acceleration (see "Anthology"). It makes it
tricky because the object might get "stuck" (D2 uses point physics, but
sometimes objects behave more like cubes and get stuck in dumb ways.)

D2X-XL appears to have a "stuck" detection, that simply disables
speedboost. You can use the bad case mentioned in the previous
paragraphs to circumvent forced acceleration and thus break game play.
Just move very slowly towards the entry point, then acceleration will
start while you're still partially in the previous segment, and you get
stuck. Once stuck detection disables acceleration, you can move freely
within the speedboost area.

D2X-XL has a SEGMENT_FUNC_SPEEDBOOST segment type, but I haven't seen a
level that uses it. It appears to be a hack to prevent "escaping" from a
speedboost area by using a path that doesn't have a disable trigger (I
think).

Yet another problem is that you can't recover items dropped into
speedboost areas. They're not getting accelerated.

Considering all these problems, and the fact that I've found no levels
that use it in a way that use it in a way that's more complex than a
tube (even if it can be entered sideways, or has ways to escape), I'm
going for the most primitive option.

Simply implement it as additional thrust vector (the thing used for user
input), that is set or unset when passing the triggers. Instead of stuck
detection or anything like this, just allow the user to move a bit. In
theory you'd need to support acceleration for robots, but I don't know a
level which actually needs this (or even just allows to observe this).
There may be places where it needs to be reset manually (such as
teleportation), but not bothering with it until it breaks.