state: save current speedboost, add crude code for extension mechanism
It's another single value of state to save.

Bumping the version is getting annoying, and the idea with using
different end markers to achieve forward-compatible extension (that
END_MAGIC23 thing) was annoying and impractical. So add a
type/length/value concept to the end of the file, that tries to be
clever about forward-compatibility. (Basically all non-trivial file
formats use this concept.) The new state to save is wrapped into such a
record.

Not sure yet if this will be useful, or if I'll regret it later. Since
the compatibility handling isn't actually tested (which would be
required to make this useful for bisecting bugs while using savegames),
so maybe the former will happen.