gameseq: remove secret level death code duplication
The code for dying in a secret level was obviously duplicated. The first
instance of the code handles dying when the reactor countdown reaches 0,
the second instance is for normal death. In both cases, the game returns
to the normal level (or advances to the next level if it's destroyed),
with some minor differences. Someone at Parallax (I guess?) didn't
bother with writing proper code and just copy&pasted this case. Meh.

Remove the duplication. This shouldn't change any behavior.