gameseq: remove some more secret exit code duplication
The previous commit reduces some duplication. But the normal secret exit
code (though an exit, i.e. a switch.c trigger) was still duplicated.

Unlike the previous commit, this commit explicitly changes behavior. It
appears the existing code did not save the secret level properly if the
base level was destroyed, and you're advancing to the next level. This
is news to me (but the previous commit did not change this). I won't
bisect it now (the absence of debug mode makes this annoying), but I
doubt I broke this myself. Whatever, fix it.

Reproduction:
- start level 5
- hit DEL+Y (countdown sequence starts)
- go through secret exit
- die in secret level (but don't destroy secret level)
- continue playing until level 8
- in level 8, it'll say the secret level got destroyed

This is because SECRETB_FILENAME doesn't exist anymore (base level was
destroyed), and on this code paths the secret level is never saved. This
is easily fixed by always saving the secret level, unless it really gets
destroyed.

For this reason, also move the code from cntrlcen.c to gameseq.c, so the
SECRETC file always gets deleted. (In theory this might make a
difference with d2x-xl, but since we don't save the level at all, it
won't make a useful difference.)

For now, just keep some strange stuff, like weapon selection backup
(fairly sure this should be done elsewhere, if it even does anything,
but let's not touch it unnecessarily).