gameseq, state: allow saving in secret levels
This is surprisingly simple. The trick to avoid extra complexity on
loading is to make the "main" file the secret level. You can just load
it, and the normal level is copied from the usual secret state file
(<N>secret.sgc) to secret.sgb. You simply save and restore the state as
it is at runtime, and done.

One tricky issue is that the game could keep state regarding the main
level simply in memory, without ever writing it to disk. But I don't
think there's much such state. I only found Entered_from_level, for
which this commit adds saving/restoring (using a messy hack by reusing
the now unused between_levels field). There could be state I overlooked.

The hardest part is making sure that everything works. This will happen
over the course of the next weeks, months, and years.

Should this feature exist? Technically, it seems to be no problem, so
here we go. For players and gameplay, it creates an unfair advantage
over the original game. Maybe reloading a secret level should set the
cheat flag?