player, state, cfile: stop using entire struct player for I/O
Instead of writing/reading struct player as a memory dump, write/read
individual fields. This makes it easier to extend the struct, and
doesn't require weird hacks to keep proper alignment.

This is cleaner in general, but is also needed for the next commit in
order to extend the player struct without breaking old savegames.

Also add a horrible new concept for simplifying I/O. The sd_* functions
(well, macros) handle both serialization and deserialization, so there's
no need to duplicate saving and loading functions. It's also pretty
close to an ideal and minimal mechanism. (With any such a system, there
needs to be a place where you define the type and order of the binary
data in the file. You still need a different struct, because the order
in memory may be different and so on.) An ideal mechanism would have
the file format in a declarative way, though. I considered doing so,
but I think it'd be too much of a pain to keep a "magic" serializer
extensible, since you'd need all the compatibility hacks in a
declarative specification.

This could also in theory handle endian swapping, since the type of each
data item is known. This is a theoretical thing since the old game
didn't give a shit about endian; however it simply used endian swapped
data files for Mac support (BE PowerPC), so this may become important
again when reading data files.

Instead of CFILE_var_w/CFILE_var_r, which I introduced earlier and which
were used in the trigger savegame code, I'm explicitly specifying the
type in the sd_* function names. This is to get some minor type safety
(changing a struct type will result in a compiler warning). Not making
the type a macro argument (like sd_var(type, sd, ptr)) may help slightly
iff endian swapping is needed later. In theory, you could also just use
_Generic to dispatch these things, but I'd rather not.