Remove incompatible legacy SDL API usage
This is from SDL1 times. It likely doesn't do anything in SDL2. SDL2
uses the hint API instead of requiring the API user to mutate the
environment variables. Along with that, they removed the SDL_putenv
wrapper, so this code doesn't compile anymore.

I didn't find an equivalent hint for this LOCK_KEYS stuff, the
enviornment variable doesn't appear in the SDL2 source code, and I have
no idea what it did. So just remove this without replacement. (Maybe the
entire option should be dropped - too lazy for this right now.)

I'm not sure how this built before. SDL_putenv seems to have been gone
for quite a while.