gamesave: do not load linked_wall field from levels
It turns out that linked_wall is never needed from levels, and some
D2X-XL levels (Anthology) have in fact blatantly wrong values set.

Not loading the field fixes some extremely weird and buggy door behavior
in the aforementioned level. Opening a door could also open an unrelated
close by door, and sometimes a door would remain stuck in the opened
state, while being rendered in the closed state.

The reason was that linked_wall was set to the wall of an apparently
unrelated door. This confused the door logic (rain of assertion
failures). In particular, it manifested when opening the mislinked doors
at the same time.

We never need linked_wall. It's only set for active doors (open doors,
or doors in progress of being opened or closed), and thus part of the
savegame state. But active doors are not loaded from levels as written
by level editors. (There is code for it, but it has been commented out
since ancient times. It was active in the original game, but commented
in all of d2x svn, which is a WTF in itself.) All of the official levels
always have this field set to -1.

Assume that if the field is set, it's just a bug. Warn against it,
because I want to know when it happens.