gamecntl: "cheat" to start up d2x-xl equipment centers
It's not clear whether d2x-xl levels require equipment centers to be
active or disabled at the start. There are definitely a number of levels
that have these, but no trigger to enable them. As a compromise, add a
feature to enable them as a pseudo-cheat (with no cheat penalty).