collide/switch: fix potentially untriggered switches on blasting panels
Sometimes, panel-switches were blown up, without triggering the
associated switch. This shouldn't happen and could lead to stuck
gameplay. This very rarely happened to me, but when it did, it was not
very debugable.

I suspect this is the cause. An object, such as a weapon fire, has a
link to its creator via parent_num. Nothing prevents that the parent
gets destroyed, and parent_num gets replaced by another object, breaking
game logic. In particular, the collision code and the switch code had
duplicated checks for the same thing (whether the parent was player or
guidebot), which were subtly different, and thus could have created this
issue. The check in collide.c could have determined that it's ok to blow
up the switch, while the check in switch.c determined that it was not
ok, thus not running the trigger, even though the switch was destroyed.

Try to fix this by removing the duplicated check from switch.c, and
improving the one in switch.c. The "improvement" is not that important,
since the disagreement (leading to unfinishable levels) cannot happen
anymore.

I'm not sure if this is really the actual cause. I think this would
require that the player was destroyed while the switch was hit (?),
which I'm not sure about.