switch: fix messed up triggers when changing levels
This fixes broken triggers in certain levels. The previous commit was a
speculative fix for the same issue, although it wasn't the actual cause
after all.

It appears that if a TT_EXIT trigger is run, the new level will be
loaded in a sub-function, so you get TONS of reentrancy problems. Commit
d0a025d4a602f187ce57b added a new one when it refactored the trigger
code in a seemingly innocuous manner.

In this case, the new level simply had a TF_ONE_SHOT trigger in the same
array slot as the TT_EXIT trigger, which meant it was disabled when the
trigger function returned after the new level was loaded. This happened
in "Fight For Your Life" levels 2 and 11 (when using the exit on levels
1 and 10).

Fix this by returning immediately if an exit was used. (In theory, it
might be possible that someone tries to use TF_ONE_SHOT with an exit,
but I've never seen this in practice.)

A cleaner fix is doing the level switch some time later in the game's
main loop. But that could cause other problems.